IOT-1836 Make resource/csdk/security/provisioning W4 compliant.
These changes include the following:
- Fix W4 warnings under resource/csdk/security/provisioning

Other changes to resource/csdk/security/ are tracked in the IOT-1835 code review.

Reviewed-on: https://gerrit.iotivity.org/gerrit/17401
Tested-by: jenkins-iotivity <jenkins@iotivity.org>

+++ resource/csdk/security/unittest/aclresourcetest.cpp
@@ -137,7 +137,7 @@ TEST(ACLResourceTest, CBORDefaultACLConversion)
     OCStackResult convRet = AclToCBORPayload(defaultAcl, &defaultPsStorage, &defaultAclSize);
     EXPECT_EQ(OC_STACK_OK, convRet);
     ASSERT_TRUE(NULL != defaultPsStorage);
-    EXPECT_NE(0, defaultAclSize);
+    EXPECT_NE(static_cast<size_t>(0), defaultAclSize);
 
     OicSecAcl_t* convertedAcl = CBORPayloadToAcl(defaultPsStorage, defaultAclSize);
     ASSERT_TRUE(NULL != convertedAcl);
